apply shared error handling to server
