+++ src/core/errorHandler.js
@@ -86,31 +86,3 @@ export function withErrorHandling({ name, id } = {}) {
     )(ErrorHandlerComponent);
   };
 }
-
-const errorPageText = {
-  401: 'Unauthorized',
-  404: 'Not Found',
-  500: 'Internal Server Error',
-};
-
-export function getErrorMsg(statusCode) {
-  const statusKey = statusCode.toString();
-  return errorPageText[statusKey];
-}
-
-export function getReduxConnectError(reduxConnectLoadState) {
-  // Create a list of any apiErrors detected.
-  const apiErrors = Object.keys(reduxConnectLoadState)
-    .map((item) => reduxConnectLoadState[item].error);
-
-  if (apiErrors.length === 1) {
-    // If we have a single API error reflect that in the page's response.
-    const apiStatus = apiErrors[0].response.status;
-    return getErrorMsg(apiStatus);
-  } else if (apiErrors.length > 1) {
-    // Otherwise we have multiple api errors it should be logged
-    // and throw a 500.
-    log.error(apiErrors);
-    return getErrorMsg(500);
-  }
-}
